requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
exceptiongroup==1.1.3
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-auth==2.23.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.1
openai==1.1.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.25.0
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.11.3
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
streamlit==1.28.1
streamlit-feedback==0.1.2
style==1.1.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trubrics==1.6.2
typeguard==2.13.3
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
update==0.0.1
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.17.0
